Refactoring
Refactoring code and putting it in their respective classes
