+++ go.mod
@@ -3,12 +3,12 @@ module github.com/airidasz/SemestroProjektasBackendKTU2020
 go 1.13
 
 require (
-	github.com/go-sql-driver/mysql v1.5.0 // indirect
-	github.com/gofrs/uuid v3.2.0+incompatible // indirect
+	github.com/go-sql-driver/mysql v1.5.0
+	github.com/gofrs/uuid v3.2.0+incompatible
 	github.com/gorilla/context v1.1.1 // indirect
-	github.com/gorilla/mux v1.7.4 // indirect
+	github.com/gorilla/mux v1.7.4
 	github.com/gorilla/sessions v1.2.0 // indirect
-	github.com/jinzhu/gorm v1.9.12 // indirect
-	github.com/wader/gormstore v0.0.0-20190904144442-d36772af4310 // indirect
-	golang.org/x/crypto v0.0.0-20200302210943-78000ba7a073 // indirect
+	github.com/jinzhu/gorm v1.9.12
+	github.com/wader/gormstore v0.0.0-20190904144442-d36772af4310
+	golang.org/x/crypto v0.0.0-20200302210943-78000ba7a073
 )
